+++ BOCCHI.World/Data/CriticalEncounterFactory.cs
@@ -1,6 +1,7 @@
 ﻿using BOCCHI.Common.Data.CriticalEncounters;
 using BOCCHI.Common.Data.Zones;
 using FFXIVClientStructs.FFXIV.Client.Game.InstanceContent;
+using System.Numerics;
 
 namespace BOCCHI.CriticalEncounters.Data;
 
@@ -14,8 +15,12 @@ public class CriticalEncounterFactory(IZoneProvider zones) : ICriticalEncounterF
     public CriticalEncounter Create(DynamicEvent ev)
     {
         CriticalEncounterId id = new(ev.DynamicEventId);
-        float radius = zones.GetZone().GetCriticalEncounterRadius(ev.DynamicEventId);
+        IZone zone = zones.GetZone();
+        float radius = zone.GetCriticalEncounterRadius(ev.DynamicEventId);
+        Vector3 fallback = zone.GetCriticalEncounterData()
+            .FirstOrDefault(a => a.Id == ev.DynamicEventId)
+            ?.Position ?? Vector3.NaN;
 
-        return new(id, ev, radius);
+        return new(id, ev, radius, fallback);
     }
 }